pyproject.toml
[tool.poetry]
name = "dobrokod-further"
version = "0.1.0"
description = ""
authors = ["werserk <werserk@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.30.1"
requests = "^2.31.0"
telebot = "^0.0.5"
python-dotenv = "^1.0.1"
streamlit = "^1.34.0"
sqlalchemy = "^2.0.30"
flask = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
